chore(storybook): convert .storybook config to TS, framework=react-vite
- main.ts uses @storybook/react-vite framework
- preview.ts uses CSF3 Preview type with named export
- manager.ts imports from storybook/manager-api
- Delete .storybook/main.js, preview.js, manager.js, webpack.config.js
- Rename stories/reactist explainers + KeyCapturer .stories.mdx -> bare .mdx
- Set typescript.reactDocgen = false to work around storybook 9.1.20 bug
  with bundled find-up@7 (ESM) and CJS preset code

React Compiler integration via @vitejs/plugin-react is deferred to follow-up;
storybook builds run without compiler. The library build (rollup) still runs
the compiler. Verify storybook build succeeds (182 indexed entries).
