+++ stories/components/KeyCapturer.mdx
@@ -1,5 +1,5 @@
 import { Story, Meta } from '@storybook/addon-docs'
-import { action } from '@storybook/addon-actions'
+import { action } from 'storybook/actions'
 
 import KeyCapturer from '../../src/components/key-capturer'
